+++ ml3d/datasets/utils/dataprocessing.py
@@ -6,6 +6,7 @@ from os.path import exists, join, isfile, dirname, abspath, split
 from open3d.ml.contrib import subsample
 from open3d.ml.contrib import knn_search
 
+from .operations import *
 
 class DataProcessing:
 
@@ -164,164 +165,6 @@ class DataProcessing:
         return np.expand_dims(ce_label_weight, axis=0)
 
     @staticmethod
-    def projection_matrix_to_CRT_kitti(proj):
-        """Split projection matrix of kitti.
-        P = C @ [R|T]
-        C is upper triangular matrix, so we need to inverse CR and use QR
-        stable for all kitti camera projection matrix.
-        Args:
-            proj (p.array, shape=[4, 4]): Intrinsics of camera.
-        Returns:
-            tuple[np.ndarray]: Splited matrix of C, R and T.
-        """
-
-        CR = proj[0:3, 0:3]
-        CT = proj[0:3, 3]
-        RinvCinv = np.linalg.inv(CR)
-        Rinv, Cinv = np.linalg.qr(RinvCinv)
-        C = np.linalg.inv(Cinv)
-        R = np.linalg.inv(Rinv)
-        T = Cinv @ CT
-        return C, R, T
-
-    @staticmethod
-    def get_frustum(bbox_image, C, near_clip=0.001, far_clip=100):
-        """Get frustum corners in camera coordinates.
-        Args:
-            bbox_image (list[int]): box in image coordinates.
-            C (np.ndarray): Intrinsics.
-            near_clip (float): Nearest distance of frustum.
-            far_clip (float): Farthest distance of frustum.
-        Returns:
-            np.ndarray, shape=[8, 3]: coordinates of frustum corners.
-        """
-        fku = C[0, 0]
-        fkv = -C[1, 1]
-        u0v0 = C[0:2, 2]
-        z_points = np.array([near_clip] * 4 + [far_clip] * 4,
-                            dtype=C.dtype)[:, np.newaxis]
-        b = bbox_image
-        box_corners = np.array(
-            [[b[0], b[1]], [b[0], b[3]], [b[2], b[3]], [b[2], b[1]]],
-            dtype=C.dtype)
-        near_box_corners = (box_corners - u0v0) / np.array(
-            [fku / near_clip, -fkv / near_clip], dtype=C.dtype)
-        far_box_corners = (box_corners - u0v0) / np.array(
-            [fku / far_clip, -fkv / far_clip], dtype=C.dtype)
-        ret_xy = np.concatenate([near_box_corners, far_box_corners],
-                                axis=0)  # [8, 2]
-        ret_xyz = np.concatenate([ret_xy, z_points], axis=1)
-        return ret_xyz
-
-    @staticmethod
-    def camera_to_lidar(points, r_rect, velo2cam):
-        """Convert points in camera coordinate to lidar coordinate.
-        Args:
-            points (np.ndarray, shape=[N, 3]): Points in camera coordinate.
-            r_rect (np.ndarray, shape=[4, 4]): Matrix to project points in
-                specific camera coordinate (e.g. CAM2) to CAM0.
-            velo2cam (np.ndarray, shape=[4, 4]): Matrix to project points in
-                camera coordinate to lidar coordinate.
-        Returns:
-            np.ndarray, shape=[N, 3]: Points in lidar coordinate.
-        """
-        points_shape = list(points.shape[0:-1])
-        if points.shape[-1] == 3:
-            points = np.concatenate(
-                [points, np.ones(points_shape + [1])], axis=-1)
-        lidar_points = points @ np.linalg.inv((r_rect @ velo2cam).T)
-        return lidar_points[..., :3]
-
-    @staticmethod
-    def corner_to_surfaces_3d(corners):
-        """Convert 3d box corners from corner function above to surfaces that
-        normal vectors all direct to internal.
-        Args:
-            corners (np.ndarray): 3d box corners with the shape of (N, 8, 3).
-        Returns:
-            np.ndarray: Surfaces with the shape of (N, 6, 4, 3).
-        """
-        # box_corners: [N, 8, 3], must from corner functions in this module
-        num_boxes = corners.shape[0]
-        surfaces = np.zeros((num_boxes, 6, 4, 3), dtype=corners.dtype)
-        corner_idxes = np.array([
-            0, 1, 2, 3, 7, 6, 5, 4, 0, 3, 7, 4, 1, 5, 6, 2, 0, 4, 5, 1, 3, 2, 6,
-            7
-        ]).reshape(6, 4)
-        for i in range(num_boxes):
-            for j in range(6):
-                for k in range(4):
-                    surfaces[i, j, k] = corners[i, corner_idxes[j, k]]
-        return surfaces
-
-    @staticmethod
-    def surface_equ_3d(polygon_surfaces):
-        """
-        Args:
-            polygon_surfaces (np.ndarray): Polygon surfaces with shape of
-                [num_polygon, max_num_surfaces, max_num_points_of_surface, 3].
-                All surfaces' normal vector must direct to internal.
-                Max_num_points_of_surface must at least 3.
-        Returns:
-            tuple: normal vector and its direction.
-        """
-        # return [a, b, c], d in ax+by+cz+d=0
-        # polygon_surfaces: [num_polygon, num_surfaces, num_points_of_polygon, 3]
-        surface_vec = polygon_surfaces[:, :, :2, :] - \
-            polygon_surfaces[:, :, 1:3, :]
-        # normal_vec: [..., 3]
-        normal_vec = np.cross(surface_vec[:, :, 0, :], surface_vec[:, :, 1, :])
-        # print(normal_vec.shape, points[..., 0, :].shape)
-        # d = -np.inner(normal_vec, points[..., 0, :])
-        d = np.einsum('aij, aij->ai', normal_vec, polygon_surfaces[:, :, 0, :])
-        return normal_vec, -d
-
-    @staticmethod
-    def points_in_convex_polygon_3d(points,
-                                    polygon_surfaces,
-                                    num_surfaces=None):
-        """Check points is in 3d convex polygons.
-        Args:
-            points (np.ndarray): Input points with shape of (num_points, 3).
-            polygon_surfaces (np.ndarray): Polygon surfaces with shape of \
-                (num_polygon, max_num_surfaces, max_num_points_of_surface, 3). \
-                All surfaces' normal vector must direct to internal. \
-                Max_num_points_of_surface must at least 3.
-            num_surfaces (np.ndarray): Number of surfaces a polygon contains \
-                shape of (num_polygon).
-        Returns:
-            np.ndarray: Result matrix with the shape of [num_points, num_polygon].
-        """
-        max_num_surfaces, max_num_points_of_surface = polygon_surfaces.shape[
-            1:3]
-        # num_points = points.shape[0]
-        num_polygons = polygon_surfaces.shape[0]
-        if num_surfaces is None:
-            num_surfaces = np.full((num_polygons,), 9999999, dtype=np.int64)
-        normal_vec, d = DataProcessing.surface_equ_3d(
-            polygon_surfaces[:, :, :3, :])
-        # normal_vec: [num_polygon, max_num_surfaces, 3]
-        # d: [num_polygon, max_num_surfaces]
-        max_num_surfaces, max_num_points_of_surface = polygon_surfaces.shape[
-            1:3]
-        num_points = points.shape[0]
-        num_polygons = polygon_surfaces.shape[0]
-        ret = np.ones((num_points, num_polygons), dtype=np.bool_)
-        sign = 0.0
-        for i in range(num_points):
-            for j in range(num_polygons):
-                for k in range(max_num_surfaces):
-                    if k > num_surfaces[j]:
-                        break
-                    sign = (points[i, 0] * normal_vec[j, k, 0] +
-                            points[i, 1] * normal_vec[j, k, 1] +
-                            points[i, 2] * normal_vec[j, k, 2] + d[j, k])
-                    if sign >= 0:
-                        ret[i, j] = False
-                        break
-        return ret
-
-    @staticmethod
     def remove_outside_points(points, rect, Trv2c, P2, image_shape):
         """Remove points which are outside of image.
         Args:
@@ -335,15 +178,16 @@ class DataProcessing:
         Returns:
             np.ndarray, shape=[N, 3+dims]: Filtered points.
         """
-        C, R, T = DataProcessing.projection_matrix_to_CRT_kitti(P2)
+        C, R, T = projection_matrix_to_CRT_kitti(P2)
         image_bbox = [0, 0, image_shape[1], image_shape[0]]
-        frustum = DataProcessing.get_frustum(image_bbox, C)
+        frustum = get_frustum(image_bbox, C)
         frustum -= T
         frustum = np.linalg.inv(R) @ frustum.T
-        frustum = DataProcessing.camera_to_lidar(frustum.T, rect, Trv2c)
-        frustum_surfaces = DataProcessing.corner_to_surfaces_3d(
+        frustum = camera_to_lidar(frustum.T, rect, Trv2c)
+        frustum_surfaces = corner_to_surfaces_3d(
             frustum[np.newaxis, ...])
-        indices = DataProcessing.points_in_convex_polygon_3d(
+        indices = points_in_convex_polygon_3d_jit(
             points[:, :3], frustum_surfaces)
         points = points[indices.reshape([-1])]
+
         return points
